* fixed issue when sending location reminder for current day

+++ src/Service/Booking.php
@@ -134,7 +134,7 @@ class Booking {
 			return;
 		}
 
-		$daysBeforeStart = Settings::getOption( 'commonsbooking_options_reminder', 'booking-'.$type.'-location-reminder-day' );
+		$daysBeforeStart = (int) Settings::getOption( 'commonsbooking_options_reminder', 'booking-'.$type.'-location-reminder-day' );
 		$startDate = strtotime( '+' . $daysBeforeStart . ' days midnight' );
 
 		// start day of booking at 23:59
